Allow empty values (null, undefined, etc)
We had a problem where titleize was passed in a null value from the database and it threw an error.

This is the only helper that will explicitly throw and error and stop the page from loading if the value isn't right. Another suggested changed would be 

```
if (typeof string !== 'string') {
  Ember.assert(`Expected a string, got a ${typeof string}`);
}
```

So an invalid value doesn't kill the page.